Add a gradient descent method
The idea of this test is somewhat similar to that of TestList_LineFit.
Instead of assuming that our data fits a linear line between our min
and max, we sample around min and max (1 point near each) to get the
local gradient. Once we have that, we calculate a linear derivative of
the line that approximates the endpoints' locations and the tangent
line at each. From there, we propagate up y-intercept points and plug
them into the inverse function of our line to get an estimate.

The method seems very competitive. While it loses out to LinearFit on
some data distributions (linear and random), it appears to only be by a
constant value. On all other distributions, it comes out on top by a
wide margin.